# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/premium": {
            "get": {
                "description": "This API List all premiums",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Premium"
                ],
                "summary": "This API List all premiums",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.FPPremium"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    }
                }
            }
        },
        "/premium/add": {
            "post": {
                "description": "This API Creates a new premium",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Premium"
                ],
                "summary": "This API Creates a new premium",
                "parameters": [
                    {
                        "description": "New Premium",
                        "name": "premium",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.FPCreatePremium"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FPPremium"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    }
                }
            }
        },
        "/premium/delete/{id}/{spId}": {
            "delete": {
                "description": "This API Delete premium by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Premium"
                ],
                "summary": "This API Delete premium by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The id of FPPremium",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The id of ServiceProvider",
                        "name": "spId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    }
                }
            }
        },
        "/premium/find/{id}/{spId}": {
            "get": {
                "description": "This API Finds premium by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Premium"
                ],
                "summary": "This API Finds premium by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The id of FPPremium",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The id of FPPremiumServiceProvider",
                        "name": "spId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FPPremium"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    }
                }
            }
        },
        "/premium/findbysp/{spId}": {
            "get": {
                "description": "This API Finds premium by ServiceProviderID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Premium"
                ],
                "summary": "This API Finds premium by ServiceProviderID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The id of FPPremiumServiceProvider",
                        "name": "spId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.FPPremium"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    }
                }
            }
        },
        "/premium/update": {
            "put": {
                "description": "This API Update existing premium",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Premium"
                ],
                "summary": "This API Update existing premium",
                "parameters": [
                    {
                        "description": "Update Premium",
                        "name": "premium",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.FPUpdatePremium"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FPPremium"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.FPAPIResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "dto.FPCreatePremium": {
            "type": "object",
            "required": [
                "fp_premium_amount",
                "fp_premium_description",
                "fp_premium_sp_id",
                "fp_premium_status"
            ],
            "properties": {
                "fp_premium_amount": {
                    "type": "string"
                },
                "fp_premium_description": {
                    "type": "string"
                },
                "fp_premium_sp_id": {
                    "type": "string"
                },
                "fp_premium_status": {
                    "type": "string"
                }
            }
        },
        "dto.FPUpdatePremium": {
            "type": "object",
            "required": [
                "fp_premium_amount",
                "fp_premium_description",
                "fp_premium_id",
                "fp_premium_sp_id",
                "fp_premium_status"
            ],
            "properties": {
                "fp_premium_amount": {
                    "type": "string"
                },
                "fp_premium_description": {
                    "type": "string"
                },
                "fp_premium_id": {
                    "type": "string"
                },
                "fp_premium_sp_id": {
                    "type": "string"
                },
                "fp_premium_status": {
                    "type": "string"
                }
            }
        },
...
        "models.FPPremium": {
            "type": "object",
            "properties": {
                "fppremiumDescription": {
                    "type": "string"
                },
                "fppremiumID": {
                    "type": "string"
                },
                "fppremiumName": {
                    "type": "string"
                },
                "fppremiumServiceID": {
                    "type": "string"
                },
                "fppremiumStatus": {
                    "type": "string"
                }
            }
        },
...
    }
}